Abstract Html.Events.Extra.Mouse so I can use metaKey
yeah, seriously. I had to re-implement half of this library so I could capture metaKey in clicks. Maybe there's a neater way to do this but I'm not sure what it is yet.

In order to add metaKey to the Keys object, I had to redefine the keys decoder, and I wanted my Event to behave like the original one, which needs all the original decoders, which aren't exposed, and it kinda snowballed.